hashmap reftrace added
